pyproject.toml
# pyproject.toml

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "github-manager-cli"
description = "A tool to manage GitHub objects"
dynamic = ["version"]
authors = [{ name = "Slav", email = "sosadtsia@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: GNU Library or Lesser General Public License (LGPL)",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Utilities",
]
keywords = ["github", "cli"]
requires-python = ">=3.11"

dependencies = [
    "PyGithub",
    "python-dotenv",
    "PyYAML",
    "rich",
    "requests"

]

[project.optional-dependencies]
dev = ["uv", "black", "pip-tools", "pytest"]

[project.urls]
Homepage = "https://github.com/SLAVNG/github-manager-cli"

[tool.pytest.ini_options]
python_files = "tests/*.py"
addopts = "--color=yes"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
path = "src/ghrm/__version__.py"
pattern = "VERSION = \"(?P<version>[^\"]+)\""

[tool.hatch.build.targets.wheel]
packages = ["src/ghrm"]

[project.scripts]
ghrm = "ghrm.cli:run_cli"
